Prefer 64-bit powershell
Previously, the powershell provider would always use the 32-bit
powershell.exe  when run on a 64-bit Windows OS, which isn't desirable.

Ideally, puppet's internal `Puppet::Util.which` call handle this (giving
preference to c:\windows\sysnative before c:\windows\system32 on x64), but
it doesn't. See #12980.

So for now, we handle this in the module itself.