Write script content to a tempfile
Previously, we were passing the powershell script to execute on the command
line. This is undesirable because the script contents are available for any
other process to see. Also, CreateProcess has a maximum command line length
of 32K characters.

I would have liked to create a tempfile and execute powershell with the
-File <path> option. However, powershell requires that puppet close the
file handle (or resort to native Win32 APIs to open the file with full
sharing). The former would expose puppet to TOCTOU race conditions. The
latter is hard, since ruby is the one creating the tempfile.

This commit executes powershell with redirected stdin. This solves the
original issues with passing the script on the commandline. And since
puppet never closes the tempfile handle until after powershell exits, the
file can't be modified after puppet writes the tempfile, and before
powershell writes it.

Note current versions of Windows use per-user temp directories with strong
permissions to minimize the likelihood of TOCTOU attacks on tempfiles, but
I'd rather not make (poor) assumptions, especially on Windows 2003.