(MODULES-2207) Fix Specs
The specs have been broken since 3.4.0 was released, because the specs
were originally testing internal Puppet code that was changed with this commit:
https://github.com/puppetlabs/puppet/commit/026b2f7ea996bb3ab5acbc0214919f3d5bc0754f.

The decision to test that far in may have been an easier decision point at the time
considering how much more difficult it has been getting rspec/mocha to evaluate
properly against a call to a super method within the confines of Puppet types and
providers. It has involved a combination of stubbing any calls to
`Puppet::Provider::Exec.any_instance`, expecting against calls to
`Puppet::Type::Exec::ProviderPowerShell.any_instance`, and aliasing the `run` method
in the spec to `run_spec_override` to allow calling the overridden method first and
only evaluating the call to the super method.

As an added bonus, all but 3 specs will run on non-Windows platforms.