(MODUlES-2962) Reuse PowerShell Session
Port of the PowerShell session code from the DSC module. This code
includes some significant speed and system utilization improvements.
In the prior version every call to exec started a brand new PowerShell
process for each of the exec modes: command, onlyif and unless. This
was costly in both time and system resources as PowerShell takes
significant time to intialize and frequent process creations spike cpu
on the target system.

In this version, a single PowerShell process is created that lives not
only for all the modes of exec, but also for the entirety of the Puppet
run. We still have a one time cost of time and system utilization for
process creation for the entire run, but gain the ability to run
consecutive commands.

A second improvement to this version is that we directly pass commands
to PowerShell through shell redirection, and not through reading a file
in like the last version. This avoids any of the parsing errors
(MODULES-2634) we have seen. This also improves syntax checking, as we
now validate the passed in commands using PowerShell itself before
the command is run, allowing us to catch errors sooner and report them.