(MODULES-3192) Use file() instead of template()
Prior to this, the README had examples of containing multi-line or
complex PowerShell scripts to be `exec`'d as templates. The examples
shown weren't actually templates and didn't have any ERB code in them.
This led to some confusion on first glance because the PowerShell code
shown didn't have any dynamic code in it; it was static code. It also
could lead some users to believe that the template() function is required.

This commit fixes that by converting the examples shown to using the
file() function and static files. It also leaves a note that actual
templates can be used if the PowerShell scripts do in fact need to be
dynamic or access Puppet variables.