(MODULES-2634) Verify try/catch with PowerShell module
Due to the changes from MODULES-2962, ticket MODULES-2634 is fixed
as a side effect. Besides the module itself using try/catch semantics
in a templated PS1 file, it supports any kind of syntax or formatting.

This commit adds a unit test and an acceptance test to ensure that this
is the case.