(MODULES-3347) Handle PS Verbose Output in Tests
When the PowerShell Informational Output Streams were implemented in
MODULES-3137 tests were added to ensure each output stream produced
the correct text. The Verbose test proved brittle because we forced
$VerbosePreference to Continue, which caused extra verbose messages
to be written because of PowerShell module auto loading.

This fixes the test to look for our text instead of requiring only
our text