(MODULES-3565) Change the working directory if specified in the resource
Previously  the cwd resource parameter was not being used during powershell
script execution.  This commit checks to see if the cwd parameter is specified
on the resource and if so, modifies the `SessionStateProxy.Path.SetLocation` to
use the specified location, otherwise it sets the location to the original
working directory.  The original working directory the current directory of the
process which creates the powershell manager i.e. Puppet.  The working directory
is not modified across runspaces.

This commit adds tests to ensure that an error is raised if a non-existant
working directory is specified and that the working directory is set during the
powershell session.