(MODULES-3709) Respect resource timeout interval in powershell manager
Previously, the a timeout value could be specified in a resource however that
value was ignore and the default of 300s was used by the powershell manager.
This commit passes throught the timeout value from the resource into the
powershell manager and also adds an acceptance test to verify this behaviour.