(MODULES-3443) Modify error test for differences in Powershell
Previously the integration tests for testing error behavior were passing on
Powershell 3+ however the error parsing is different in Powershell 2.0. This
commit modifies the test expectations for a line and char to returned but the
content of the line and char values are ignored.  This means the tests will now
pass on all Powershell versions but still ensure the correct behavior.