(MODULES-3690) Determine Compatible PowerShell
Determine if the PowerShell version installed is compatible
with module. When folks are using PowerShell 3+, they are on .NET
Framework 4+, so they are completely compatible with the PowerShell
Manager. When folks are using PowerShell v2 and .NET 3.5, they are also
compatible, as `System.IO.Pipes` requires at least .NET Framework 3.5.

However if they are running PowerShell v2 with .NET 2.0 (which is not
going to happen very often, but it represents a base install on Windows
Server 2008 and R2), it is not compatible with the PowerShell Manager.