(MODULES-5228) Move PowerShell Manager template file
Previously the PowerShell manager template file lived in the root of the
templates directory, however so does the same file in the DSC module which has
caused issues as they diverge.  This commit moves the template file to a
PowerShell module specific location so that it should not conflict with any
other modules.