(MODULES-6081) Rename PowerShell executable
This commit renames the PowerShell binary name from PowerShell to
pwsh. This is because the name of the binary has changed in the
latest PowerShell betas. This commit also removes the manual checking
of specific paths like /usr/local/bin for the pwsh executable.
Instead for non Windows platforms we simply call pwsh and let the
platform do the path resolution based on current PATH settings. One
reason to switch to this behavior is the Mac OS X preferences
for ordering /usr/local/bin in path before /usr/bin, as some other
platforms do. The PATH checking in Windows was for specific use case
and we decided that absent that use case in other platforms, it should
be removed so that users get the behavior they expect based on their
platform, and not on internal Puppet logic.