(MODULES-7018) Fix Zero Tmeout Behavior
Some other Puppet Exec providers will treat a `timeout` value of `0` as
an infinite timeout. We have chosen not to support an infinite timeout
in this provider, so this change will ensure that a sensible value is
used if the user passes in `0` as the timeout. It also fixes the README
documentation to make it clearer that `0` is not supported, and that
attempting to use it will result in the default timeout of 300 seconds
being used instead.