Fix issues in README.md
The documentation in README.md was out of date. First, the link for the Linux installation documentation was dead (404). I tracked down the current link and used that instead. Second, the issue with a missing HOME environment variable was fixed in https://github.com/PowerShell/PowerShell/issues/1794. As a result, I have removed that section of the README.