(MODULES-8356) Search for pwsh binary
Previously the pwsh provider only looked in the system path and some
predefined paths on Windows.  However PowerShell Core can be installed
anywhere as it is now a tarball/zipfile installation.  This means a
user should be able to arbitrarily define a location for the provider
to use.

This commit removes the :pwsh command because command resolution
can not use a custom path when searching. This is normally used to see
if a provider is applicable.  It is then replaced by a method called
get_pwsh_command which uses a copy of the Puppet `which` command which
can use custom search paths when locating binaries.

Note we can not extend or update the Puppet version of this code as
the which method is private. In this case we copy the code verbatim
and then modify it.

This commit also will now fail a catalog apply if pwsh can not be
located. Previously it would error that no suitable provider could
be found.

This commit also updates the READMe and metadata.json to reflect the
operating systems that the pwsh provider will run on.