(maint) Fix module installing in master-agent scenario
Previously the powershell module was only being installed onto the
default host.  However this meant the module was only being installed
onto the agent, not the server, which then caused issues during tests
because the module content was not 'file syncd' from the master and
tests were failing as it couldn't find the powershell provider.

This commit uses the module install helper from beaker to install on
all hosts.