(MODULES-8356) Improve pwsh searching
Previously the pwsh provider was modified to search for the pwsh binary
in a user specified path.  However this introduces a lot of new code.
This commit removes the redundant code and uses a public method within
Puppet (.withenv) to actually do the searching

This commit also uses a simple memoizing pattern to remember the path
to avoid costly IO when searching.  This commit also adds tests for
this scenario.