(MODULES-8358) Refactor PowerShell Manager Execution
In preparation to add cross platform support to the PowerShell Manager,
this commit refactors the common execution code in the powershell provider into
the PuppetX::PowerShell::PowerShellManager.execute_resource method.  This
takes PowerShell code and a puppet exec resource and executes this using the
PowerShell Manager.

Previously the powershell command was a string which also contain the powershell
invocation commands.  This commit also separates the powershell command and arguments
so that we can call pwsh or powershell separately from the same PowerShell Manager.