(MODULES-8358) Modify acceptance tests for specific PowerShell Core version
Previously the acceptance tests were gated on the `powershell6` role within the
beaker hosts file.  However Beaker Host Generator doesn't understand numbers in
the role name [1].  This commit changes the PowerShell core detection to use
the key/value property `powershell` for each host.  This has the added benefit
to be able to specify the version of powershell core to be tested, not just
latest.

[1] https://github.com/puppetlabs/beaker-hostgenerator/issues/149