(MODULES-8359) Remove non-Windows support for powershell provider
Previously in MODULES-3495 cross platform support was added to the powershell
provider.  However as part of MODULES-8358 the pwsh provder was added to
achieve the same.  This means that the non-Windows platform support within the
powershell provider can be removed as it is no longer necessary.  Also note that
as this was marked as experimental, and had known issues with functioning
therefore the removal will not require a major version bump of the module.

This commit removes the support and also modifies the unit and acceptance tests
to not test the powershell provider on non-Windows platforms.