(MODULES-8748) Improve pipe reading in the PowerShell Manager
Previously the powershell manager was modified to read in data from the pipe
until all data was read, in 8K chunks.  However it was noted that this would be
causing more IO calls than needed due to the 8K chunks.

In this commit, the reader will attempt to read all bytes in a single call and
use the string concatenation method to reduce the number the potential number
of internal objects being created.

Also the is_readable? method was updated to also check that the stream is not
EOF. In odd cases the stream can show that it's ready to read, but a sysread
call will throw an EOF, for example this could occur when data is buffered in
a slow way.