(FM-8475) Replace library code
Prior to this commit the PowerShell module providers
depended on a vendored version of the PowerShell code
manager library in the PuppetX namespace.

This commit updates the providers and their tests to
leverage the ruby-pwsh gem from the pwshlib Puppet
module instead. This reduces the amount of maintenance
required and simplifies the code base.