(MODULES-10722) Inherit pipe_timeout from timeout
Prior to this commit there was no way to modify the pipe_timeout
parameter when instantiating a PowerShell Manager instance, users
were stuck with the default defined in ruby-pwsh (30s). If this
time is insufficient to start the host process, the exec failed
with no way to extend it short of modifying the provider.

This commit updates the definition of the ps_manager by checking
to see if the timeout parameter is specified on the resource and,
if so, inheriting that to re-use for the pipe timeout.